address kun's comments
